--HG--
branch : soc.2007.xmpp
rename : pidgin/pixmaps/emotes/default/24/madtongue.png => pidgin/pixmaps/emotes/default/24/mad-tongue.png
extra : convert_revision : 3901a9ccddfe4bbc05861b0351ceea85579b8a09

+++ finch/libgnt/gntwm.c
@@ -738,7 +738,7 @@ dump_screen(GntBindable *bindable, GList *null)
 			print = ch;
 #ifndef NO_WIDECHAR
 			if (wch.chars[0] > 255) {
-				snprintf(unicode, sizeof(unicode), "&#x%x;", wch.chars[0]);
+				snprintf(unicode, sizeof(unicode), "&#x%x;", (unsigned int)wch.chars[0]);
 				print = unicode;
 			}
 #endif
